requirements.txt
streamlit
pandas
openpyxl
langchain-groq
langchain-google-genai
langgraph
python-dotenv
plotly==5.20.0